+++ internal/query/options/execute.go
@@ -44,6 +44,7 @@ type (
 		retryOptions           []retry.Option
 		responsePartLimitBytes int64
 		label                  string
+		isImplicitSession      bool
 	}
 
 	// Execute is an interface for execute method options
@@ -70,6 +71,7 @@ type (
 	}
 	execModeOption         = ExecMode
 	responsePartLimitBytes int64
+	implicitSessionOption  struct{}
 )
 
 func (poolID resourcePool) applyExecuteOption(s *executeSettings) {
@@ -190,6 +192,14 @@ func (s *executeSettings) ResponsePartLimitSizeBytes() int64 {
 
 func (s *executeSettings) Label() string {
 	return s.label
+}
+
+func (s *executeSettings) IsImplicitSession() bool {
+	return s.isImplicitSession
+}
+
+func (implicitSessionOption) applyExecuteOption(s *executeSettings) {
+	s.isImplicitSession = true
 }
 
 func WithParameters(params params.Parameters) parametersOption {