+++ internal/query/options/retry.go
@@ -109,6 +109,10 @@ func WithIdempotent() RetryOptionsOption {
 	return []retry.Option{retry.WithIdempotent(true)}
 }
 
+func WithImplicitSession() implicitSessionOption {
+	return implicitSessionOption{}
+}
+
 func WithLabel(lbl string) LabelOption {
 	return LabelOption(lbl)
 }